Create Dijkstra.java
Description: This class calculates the shortest path from one vertex to another on a specified graph, using Dijkstra's algorithm.  All possible paths are put into a PriorityQueue which is prioritized by lowest cost.  When the path with appropriate starting and ending points is first in the queue, it will be returned as the shortest path.