Add files via upload
Path objects have an ArrayList of vertices which  denote each step of the path and an ArrayList of edges that connect the vertices.  The path also has a String that contains the vertex symbols of the path. Total cost of the path is maintained by accumulating the cost of each edge in the path.